Adopt the transforms from Guotai's tensorflow code
These consist of padding and random sampling of subimages of size
128x128x32 for training and 256x128x32 for testing.

Currently, the code is for UNet3D and performs 84.5% dice score on
the test set (roughly what is expected from the paper).

The dice score for testing is now based on the loss function, which
includes a smooth term (smooth=1e-5) (same as in Guotai's code).

The learning rate now is divided by 10 (not halved as stated in the
paper) after every 100 epochs. This is also from Guotai's code.

The training time is less than one hour on a GeForce RTX 2070 Mobile.